EKIRJASTO-769 Move logo and badge components to same folder
