+++ public/locales/en/translations.json
@@ -1,4 +1,6 @@
 {
+  "badges.ariaLabelForAppStoreLink": "Download E-library on the App Store",
+  "badges.ariaLabelForGooglePlayLink": "Get E-library on Google Play",
   "book.borrowable": "Available: {{availableCopies}} / {{totalCopies}}",
   "book.differenceInDays": "{{differenceInDays}} days",
   "book.differenceInHours": "{{differenceInHours}} hours",
@@ -14,7 +16,6 @@
   "bookCover.coverOfBook": "Cover of book: ",
   "bookDetails.ariaLabelForBookDetails": "Book details",
   "bookDetails.ariaLabelForBookInfo": "Book info",
-  "bookDetails.ariaLabelForEkirjastoLogo": "Image of the E-library logo",
   "bookDetails.ariaLabelForFulfillmentCard": "Borrow and download card",
   "bookDetails.ariaLabelForRecommendations": "Recommendations",
   "bookDetails.ariaLabelForSummary": "Book summary",
@@ -137,6 +138,7 @@
   "languageSelector": "Choose language",
   "listFilters.label": "Book format",
   "loadingIndicator.loading": "Loading...",
+  "logos.ariaLabelForNatLibFiLink": "National Library of Finland",
   "magazineFilters.allMagazines": "All magazines",
   "magazineFilters.allSubjects": "All subjects",
   "magazineFilters.selectMagazine": "Select magazine",
@@ -166,7 +168,6 @@
   "myBooks.favoritesTitle": "Favorites",
   "myBooks.loansAndHoldsTitle": "Loans and Holds",
   "myBooks.myBooksTitle": "My Books",
-  "natlibfiLogos.ariaLabelForNatLibFiLink": "National Library of Finland",
   "search.applicationError": "Cannot perform search.",
   "search.ariaLabelForButton": "Search for material",
   "search.ariaLabelForTextInput": "Enter search keyword",
@@ -186,8 +187,6 @@
   "simplyELogo": "SimplyE Logo",
   "skipNavigation.skipToContent": "Skip to content",
   "statisticsPage.titleForIframe": "E-library statistics",
-  "storeBadges.ariaLabelForAppStoreLink": "Download E-library on the App Store",
-  "storeBadges.ariaLabelForGooglePlayLink": "Get E-library on Google Play",
   "useBorrow.borrowBook": "Borrow",
   "useBorrow.borrowing": "Borrowing...",
   "useBorrow.reserveBook": "Reserve",
