+++ public/locales/fi/translations.json
@@ -1,4 +1,6 @@
 {
+  "badges.ariaLabelForAppStoreLink": "Lataa E-kirjasto App Storesta",
+  "badges.ariaLabelForGooglePlayLink": "Lataa E-kirjasto Google Playsta",
   "book.borrowable": "Saatavissa: {{availableCopies}} / {{totalCopies}}",
   "book.differenceInDays": "{{differenceInDays}} päivää",
   "book.differenceInHours": "{{differenceInHours}} tuntia",
@@ -14,7 +16,6 @@
   "bookCover.coverOfBook": "Kirjan kansi:",
   "bookDetails.ariaLabelForBookDetails": "Kirjan tiedot",
   "bookDetails.ariaLabelForBookInfo": "Kirjan tiedot",
-  "bookDetails.ariaLabelForEkirjastoLogo": "E-kirjaston logon kuva",
   "bookDetails.ariaLabelForFulfillmentCard": "Laina- ja latauskortti",
   "bookDetails.ariaLabelForRecommendations": "Suositukset",
   "bookDetails.ariaLabelForSummary": "Kirjan kuvaus",
@@ -137,6 +138,7 @@
   "languageSelector": "Valitse sivuston kieli",
   "listFilters.label": "Kirjamuoto",
   "loadingIndicator.loading": "Ladataan...",
+  "logos.ariaLabelForNatLibFiLink": "Kansalliskirjasto",
   "magazineFilters.allMagazines": "Kaikki lehdet",
   "magazineFilters.allSubjects": "Kaikki aiheet",
   "magazineFilters.selectMagazine": "Valitse lehti",
@@ -166,7 +168,6 @@
   "myBooks.favoritesTitle": "Suosikit",
   "myBooks.loansAndHoldsTitle": "Lainat ja varaukset",
   "myBooks.myBooksTitle": "Kirjani",
-  "natlibfiLogos.ariaLabelForNatLibFiLink": "Kansalliskirjasto",
   "search.applicationError": "Hakua ei voi suorittaa.",
   "search.ariaLabelForButton": "Hae aineistoa",
   "search.ariaLabelForTextInput": "Syötä hakusana",
@@ -186,8 +187,6 @@
   "simplyELogo": "SimplyE-logo",
   "skipNavigation.skipToContent": "Siirry sisältöön",
   "statisticsPage.titleForIframe": "E-kirjaston tilastot",
-  "storeBadges.ariaLabelForAppStoreLink": "Lataa E-kirjasto App Storesta",
-  "storeBadges.ariaLabelForGooglePlayLink": "Lataa E-kirjasto Google Playsta",
   "useBorrow.borrowBook": "Lainaa",
   "useBorrow.borrowing": "Lainataan...",
   "useBorrow.reserveBook": "Varaa tämä kirja",
