+++ public/locales/sv/translations.json
@@ -1,4 +1,6 @@
 {
+  "badges.ariaLabelForAppStoreLink": "Hämta E-biblioteket i App Store",
+  "badges.ariaLabelForGooglePlayLink": "Ladda E-biblioteket ned på Google Play",
   "book.borrowable": "Tillgänglig: {{availableCopies}} / {{totalCopies}}",
   "book.differenceInDays": "{{differenceInDays}} dagar",
   "book.differenceInHours": "{{differenceInHours}} timmar",
@@ -14,7 +16,6 @@
   "bookCover.coverOfBook": "Bokomslag:",
   "bookDetails.ariaLabelForBookDetails": "Bokinformation",
   "bookDetails.ariaLabelForBookInfo": "Bokinformation",
-  "bookDetails.ariaLabelForEkirjastoLogo": "Bild på E-bibliotekets logotyp",
   "bookDetails.ariaLabelForFulfillmentCard": "Låne- och nedladdningskort",
   "bookDetails.ariaLabelForRecommendations": "Rekommendationer",
   "bookDetails.ariaLabelForSummary": "Bokbeskrivning",
@@ -137,6 +138,7 @@
   "languageSelector": "Välj språk",
   "listFilters.label": "Bokformat",
   "loadingIndicator.loading": "Laddar...",
+  "logos.ariaLabelForNatLibFiLink": "Nationalbiblioteket",
   "magazineFilters.allMagazines": "Alla tidskrifter",
   "magazineFilters.allSubjects": "Alla ämnen",
   "magazineFilters.selectMagazine": "Välj tidskrift",
@@ -166,7 +168,6 @@
   "myBooks.favoritesTitle": "Favoriter",
   "myBooks.loansAndHoldsTitle": "Lån och reserveringar",
   "myBooks.myBooksTitle": "Mina böcker",
-  "natlibfiLogos.ariaLabelForNatLibFiLink": "Nationalbiblioteket",
   "search.applicationError": "Sökningen kan inte utföras.",
   "search.ariaLabelForButton": "Sök material",
   "search.ariaLabelForTextInput": "Ange sökord",
@@ -186,8 +187,6 @@
   "simplyELogo": "SimplyE-logotyp",
   "skipNavigation.skipToContent": "Hoppa till innehåll",
   "statisticsPage.titleForIframe": "E-bibliotekets statistik",
-  "storeBadges.ariaLabelForAppStoreLink": "Hämta E-biblioteket i App Store",
-  "storeBadges.ariaLabelForGooglePlayLink": "Ladda E-biblioteket ned på Google Play",
   "useBorrow.borrowBook": "Hämta",
   "useBorrow.borrowing": "Lånar...",
   "useBorrow.reserveBook": "Reservera denna bok",
