+++ src/icons/EkirjastoLogoGreenENoBackground.tsx
